agent/requirements.txt
pyirsdk>=1.0.0
python-socketio[client]>=5.9.0
